+++ django_project/core/tests/test_background_task.py
@@ -11,7 +11,6 @@ import pytz
 from ast import literal_eval as make_tuple
 from django.test import TestCase
 from django.utils import timezone
-from django.conf import settings
 
 from core.models import BackgroundTask, TaskStatus
 from core.celery import (
@@ -30,14 +29,6 @@ from core.celery import (
     conf
 )
 from core.factories import BackgroundTaskF, UserF
-
-from gap.tasks.ingestor import (
-    run_ingestor_session, notify_ingestor_failure
-)
-from gap.tasks.collector import (
-    run_collector_session, notify_collector_failure
-)
-from gap.models.ingestor import IngestorSession, CollectorSession
 
 
 mocked_dt = datetime.datetime(2024, 8, 14, 10, 10, 10, tzinfo=pytz.UTC)
@@ -345,144 +336,3 @@ class TestBackgroundTask(TestCase):
         bg_task.task_on_errors(exception="Test collector failure")
 
         mock_notify.assert_called_once_with(25, "Test collector failure")
-
-    @mock.patch("gap.tasks.ingestor.notify_ingestor_failure.delay")
-    def test_task_on_errors_ingestor_session_exception(self, mock_notify):
-        """Test triggered when IngestorSession fails with an error."""
-        bg_task = BackgroundTaskF.create(
-            task_name="ingestor_session",
-            context_id="30"
-        )
-
-        bg_task.task_on_errors(exception="Test ingestor failure")
-
-        mock_notify.assert_called_once_with(30, "Test ingestor failure")
-
-    @mock.patch("gap.tasks.ingestor.notify_ingestor_failure.delay")
-    def test_notify_ingestor_failure_session_not_found(self, mock_notify):
-        """Test when an ingestor session does not exist."""
-        with self.assertLogs("gap.tasks.ingestor", level="WARNING") as cm:
-            notify_ingestor_failure(9999, "Session not found")
-
-        self.assertIn("IngestorSession 9999 not found.", cm.output[0])
-
-    @mock.patch("gap.tasks.ingestor.IngestorSession.objects.get")
-    @mock.patch("gap.tasks.ingestor.logger")
-    def test_notify_ingestor_failure_no_admin_email(
-        self, mock_logger, mock_ingestor_get
-    ):
-        """Test when no admin emails exist."""
-        # Mock IngestorSession.objects.get to return a dummy session
-        mock_ingestor_get.return_value = mock.Mock()
-
-        self.superuser.is_superuser = False
-        self.superuser.save()
-
-        # Call function
-        notify_ingestor_failure(42, "Test failure")
-
-        # Verify log message when no admin emails exist
-        mock_logger.warning.assert_any_call(
-            "No admin email found."
-        )
-
-    @mock.patch("gap.tasks.ingestor.notify_ingestor_failure.delay")
-    @mock.patch("gap.models.ingestor.IngestorSession.objects.get")
-    def test_run_ingestor_session_not_found(self, mock_get, mock_notify):
-        """Test triggered when session is not found."""
-        mock_get.side_effect = IngestorSession.DoesNotExist
-
-        run_ingestor_session(9999)
-
-        mock_notify.assert_called_once_with(9999, "Session not found")
-
-    @mock.patch("gap.tasks.ingestor.send_mail")
-    @mock.patch("gap.tasks.ingestor.IngestorSession.objects.get")
-    def test_notify_ingestor_failure_with_admin_emails(
-        self, mock_ingestor_get, mock_send_mail
-    ):
-        """Test that email is sent when admin emails exist."""
-        # Mock IngestorSession.objects.get to return a dummy session
-        ingestor_obj = mock.Mock()
-        ingestor_obj.ingestor_type = 'test'
-        mock_ingestor_get.return_value = ingestor_obj
-
-        # Call function
-        notify_ingestor_failure(42, "Test failure")
-
-        # Ensure send_mail() was called with correct parameters
-        mock_send_mail.assert_called_once_with(
-            subject="Ingestor Failure Alert",
-            message=(
-                "Ingestor Session 42 - test has failed.\n\n"
-                "Error: Test failure\n\n"
-                "Please check the logs for more details."
-            ),
-            from_email=settings.DEFAULT_FROM_EMAIL,
-            recipient_list=["admin@example.com"],
-            fail_silently=False,
-        )
-
-    @mock.patch("gap.models.ingestor.CollectorSession.objects.get")
-    @mock.patch("gap.tasks.collector.notify_collector_failure.delay")
-    def test_run_collector_session_not_found(self, mock_notify, mock_get):
-        """Test triggered when collector session is not found."""
-        mock_get.side_effect = CollectorSession.DoesNotExist
-
-        run_collector_session(9999)
-
-        mock_notify.assert_called_once_with(
-            9999, "Collector session not found"
-        )
-
-    @mock.patch("gap.tasks.collector.notify_collector_failure.delay")
-    def test_task_on_errors_collector_session(self, mock_notify):
-        """Test triggered when collector_session fails."""
-        bg_task = BackgroundTask.objects.create(
-            task_name="collector_session",
-            context_id="15"
-        )
-        bg_task.task_on_errors(exception="Test collector failure")
-
-        mock_notify.assert_called_once_with(15, "Test collector failure")
-
-    @mock.patch("gap.tasks.collector.CollectorSession.objects.get")
-    @mock.patch("gap.tasks.collector.logger")
-    def test_notify_collector_failure_no_admin_email(
-        self, mock_logger, mock_collector_get
-    ):
-        """Test when no admin emails exist for collector failure."""
-        mock_collector_get.return_value = mock.Mock()
-
-        self.superuser.is_superuser = False
-        self.superuser.save()
-
-        notify_collector_failure(42, "Test failure")
-
-        mock_logger.warning.assert_any_call(
-            "No admin email found."
-        )
-
-    @mock.patch("gap.tasks.collector.send_mail")
-    @mock.patch("gap.tasks.collector.CollectorSession.objects.get")
-    def test_notify_collector_failure_with_admin_emails(
-        self, mock_collector_get, mock_send_mail
-    ):
-        """Test that email is sent for collector failure."""
-        ingestor_obj = mock.Mock()
-        ingestor_obj.ingestor_type = 'test'
-        mock_collector_get.return_value = ingestor_obj
-
-        notify_collector_failure(42, "Test failure")
-
-        mock_send_mail.assert_called_once_with(
-            subject="Collector Failure Alert",
-            message=(
-                "Collector Session 42 - test has failed.\n\n"
-                "Error: Test failure\n\n"
-                "Please check the logs for more details."
-            ),
-            from_email=settings.DEFAULT_FROM_EMAIL,
-            recipient_list=["admin@example.com"],
-            fail_silently=False,
-        )
